added riscv.rs for easy access to registers

+++ src/memlayout.rs
@@ -17,23 +17,23 @@
 
 // qemu puts UART registers here in physical memory.
 // IRQ: interrupt request
-pub const UART0: usize = 0x10000000;
+pub const UART0: u64 = 0x10000000;
 pub const UART0_IRQ: usize = 10;
 
 // virtio mmio interface
-pub const VIRTIO0: usize = 0x10001000;
+pub const VIRTIO0: u64 = 0x10001000;
 pub const VIRTIO0_IRQ: usize = 1;
 
 // core local interruptor (CLINT), which contains the timer.
-pub const CLINT: usize =  0x2000000;
+pub const CLINT: u64 =  0x2000000;
 // #define CLINT_MTIMECMP(hartid) (CLINT + 0x4000 + 8*(hartid))
 // - function?
-pub const CLINT_MTIME: usize = CLINT + 0xBFF8; // cycles since boot.
+pub const CLINT_MTIME: u64 = CLINT + 0xBFF8; // cycles since boot.
 
 // qemu puts platform-level interrupt controller (PLIC) here.
-pub const PLIC: usize = 0x0c000000;
-pub const PLIC_PRIORITY: usize = PLIC + 0x0;
-pub const PLIC_PENDING: usize = PLIC + 0x1000;
+pub const PLIC: u64 = 0x0c000000;
+pub const PLIC_PRIORITY: u64 = PLIC + 0x0;
+pub const PLIC_PENDING: u64 = PLIC + 0x1000;
 //#define PLIC_SENABLE(hart) (PLIC + 0x2080 + (hart)*0x100)
 //#define PLIC_SPRIORITY(hart) (PLIC + 0x201000 + (hart)*0x2000)
 //#define PLIC_SCLAIM(hart) (PLIC + 0x201004 + (hart)*0x2000)
@@ -41,8 +41,8 @@ pub const PLIC_PENDING: usize = PLIC + 0x1000;
 // the kernel expects there to be RAM
 // for use by the kernel and user pages
 // from physical address 0x80000000 to PHYSTOP.
-pub const KERNBASE: usize = 0x80000000;
-pub const PHYSTOP: usize = KERNBASE + 128*1024*1024;
+pub const KERNBASE: u64 = 0x80000000;
+pub const PHYSTOP: u64 = KERNBASE + 128*1024*1024;
 
 // map the trampoline page to the highest address,
 // in both user and kernel space.
